split the components into own files
